refactor: 기능별 파일 그룹화 user-input 체크하는 기능 state 값 관리하는 기능 result 관리하는 기능 으로 분리
