+++ src/constants/result-message.js
@@ -1,3 +1,3 @@
 export const NOTHING_MESSAGE = '낫싱';
 export const CLEAR_MESSAGE = '🎉<strong>정답을 맞추셨습니다!</strong>🎉';
-export const RESTART_MESSAGE = '게임을 새로 시작하시겠습니까? ';
+export const RESTART_MESSAGE = '<strong>게임을 새로 시작하시겠습니까? </strong>';
